extract CLRMatch into separate file
